테스트서버와 본 서버의 환경 동일하게 바꾸기

See merge request s09-final/S09P31D209!42

+++ tingting/src/main/java/com/alsif/tingting/book/controller/BookController.java
@@ -136,7 +136,7 @@ public class BookController {
 	})
 	@PostMapping("/{concertDetailSeq}/seat")
 	ResponseEntity<SuccessResponseDto> book(
-		@PathVariable("concertDetailSeq") Long concertDetailSeq, @RequestParam Long userSeq,
+		@PathVariable("concertDetailSeq") Integer concertDetailSeq, @RequestParam Integer userSeq,
 		@RequestBody ConcertSeatBookRequestDto requestDto) {
 		log.info("===== 선택 좌석 예매 요청 시작, url={}, concertDetailSeq: {}, {} =====",
 			"/concerts", concertDetailSeq, requestDto.toString());
