+++ tingting/src/main/java/com/alsif/tingting/book/repository/TicketRepository.java
@@ -21,5 +21,5 @@ public interface TicketRepository extends JpaRepository<Ticket, Integer> {
 			+ "    JOIN ConcertHall ch ON (c.concertHall.seq = ch.seq) "
 			+ "    JOIN Point p ON (p.ticket.seq = t.seq) "
 			+ "WHERE t.user.seq = :userSeq AND p.pay < 0")
-	Page<TicketBaseDto> findAllByUserSeq(@Param("userSeq") Long userSeq, Pageable pageable);
+	Page<TicketBaseDto> findAllByUserSeq(@Param("userSeq") Integer userSeq, Pageable pageable);
 }
